Move get_agent_version to tests.packaging.common
